+++ api/src/sendinblue.js
@@ -1,3 +1,4 @@
+const { captureMessage } = require("@sentry/node");
 const fetch = require("node-fetch");
 
 const { SENDINBLUEKEY, ENVIRONMENT } = require("./config");
@@ -241,12 +242,14 @@ async function sync(obj, type, { force } = { force: false }) {
 
 async function syncContact(email, attributes, listIds) {
   try {
-    const ok = await updateContact(email, { attributes, listIds });
-    if (!ok) {
+    try {
+      await updateContact(email, { attributes, listIds });
+    } catch (e) {
+      captureMessage("Contact not found during update, creating new contact:" + email);
       await createContact({ email, attributes, listIds });
     }
   } catch (e) {
-    console.log("error", e);
+    capture(e);
   }
 }